requirements.txt
numpy==1.26.4
pillow==10.3.0
scipy==1.13.0
matplotlib==3.8.4